Organize? (#10)
* moved go code into /src
* moved media to /media
* init.sql is back at /postgres/entrypoint/

Idk the best way to do the backup variables, so I just put my secrets variable back in for the backup token. Will we even be able to run this without docker once the db is active?

I was able to use the bot with both go run and docker, so I think this should be ok.